+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 * Setting names can now include the `.` character
 * Added schema information for setting types
 * Added POST, PATCH and DELETE API's to context_features, allowing run-time modifications
+* Added POST, PUT, GET and DELETE API's to edit metadata (setting and rule)
 ### Changed
 * Heksher now uses sqlalchemy instead of databases
 * inline_sql method removed, uses sqlalchemy validation instead
 * metadata is now stored in seperate tables
+* metadata keys are restricted to alphanumeric characters, underscore and dash
 ### Fixed
 * removed yellowbox as a non-dev dependency
 ## 0.3.1